Add Rect_trace; use 32-bit encoders; PID->10ms
增加矩形跟踪支持和鲁棒性，并加强电机/陀螺仪控制时序：

- 实现meet_Rect()和Rect_trace(dir, speed, speed_add)函数，以执行矩形框线跟踪。main.c现在调用Rect_trace函数，并在起点附近切换LED状态。
- 将编码器计数器/行程从int16_t扩展到int32_t以避免溢出
- 将PID控制周期从50毫秒缩短至10毫秒（在Drivers/Motor/Motor.h和mspm0-modules.syscfg文件中），以实现更快的循环更新。
- 在启用NVIC之前，通过清除UART挂起中断并清空RX FIFO来初始化陀螺仪（位于Drivers/Gyroscope/Gyroscope.c文件中），以避免接收事件失效。
- 从Motor_PID_Update_Single中移除冗长的电机PID UART调试打印。

这些更改改进了跟踪行为，防止了编码器溢出，减少了控制延迟，并使陀螺仪UART初始化更加稳健。

+++ Drivers/MSPM0/interrupt.c
@@ -5,11 +5,11 @@
 volatile uint8_t BUTTON_1_IS_PRESSED = 0;
 volatile uint8_t BUTTON_2_IS_PRESSED = 0;
 
-int16_t Motor_Left_roll = 0;
-int16_t Motor_Right_roll = 0;
+int32_t Motor_Left_roll = 0;
+int32_t Motor_Right_roll = 0;
 
-int16_t Motor_Left_Journey = 0;
-int16_t Motor_Right_Journey = 0;
+int32_t Motor_Left_Journey = 0;
+int32_t Motor_Right_Journey = 0;
 
 extern char buffer[100];
